# api/docs/swagger.yaml
definitions:
  models.ResponseError:
    properties:
      error: {}
    type: object
  models.ResponseSuccess:
    properties:
      data: {}
      metadata: {}
    type: object
...
  user_service.CreateSystemUser:
    properties:
      gmail:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
...
  user_service.GetListSystemUserRequest:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      search:
        type: string
    type: object
...
  user_service.SystemUserPrimaryKey:
    properties:
      id:
        type: string
    type: object
...
  user_service.UpdateSystemUserRequest:
    properties:
      gmail:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/system-user/create:
    post:
      consumes:
      - application/json
      description: This api creates a system-user and returns its id
      parameters:
      - description: system-user
        in: body
        name: system-user
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateSystemUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Creates a system-user
      tags:
      - SystemUser
  /v1/system-user/delete:
    delete:
      description: This api delete a system-user
      parameters:
      - description: system-user
        in: body
        name: system-user
        required: true
        schema:
          $ref: '#/definitions/user_service.SystemUserPrimaryKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: delete a system-user
      tags:
      - SystemUser
  /v1/system-user/getbyid:
    post:
      description: This api get by id a system-user
      parameters:
      - description: system-user
        in: body
        name: system-user
        required: true
        schema:
          $ref: '#/definitions/user_service.SystemUserPrimaryKey'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get by id a system-user
      tags:
      - SystemUser
  /v1/system-user/getlist:
    post:
      description: This api get list a system-user
      parameters:
      - description: system-user
        in: body
        name: system-user
        required: true
        schema:
          $ref: '#/definitions/user_service.GetListSystemUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Get list a system-user
      tags:
      - SystemUser
  /v1/system-user/update:
    put:
      consumes:
      - application/json
      description: This API updates a system-user
      parameters:
      - description: SystemUser object to update
        in: body
        name: system-user
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateSystemUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Update a system-user
      tags:
      - SystemUser
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"